+++ src/core/co_stub.cc
@@ -44,6 +44,7 @@
 #include "encoding.h"
 #include "url_parser.h"
 #include <env.h>
+#include "fvisible.h"
 
 static std::string addedScriptOnNewDocumentId = "";
 
@@ -61,7 +62,7 @@ static std::string addedScriptOnNewDocumentId = "";
  * Error Handling:
  * - If the `client_api.js` file cannot be read, an error is logged, and a message box is shown on Windows.
  */
-const std::string GetBootstrapModule(const std::vector<std::string> scriptModules, const uint16_t port)
+MILLENNIUM const std::string GetBootstrapModule(const std::vector<std::string> scriptModules, const uint16_t port)
 {
     std::string scriptModuleArray;
     std::string scriptContents = SystemIO::ReadFileSync((std::filesystem::path(GetEnv("MILLENNIUM__SHIMS_PATH")) / "client_api.js").string());
@@ -95,7 +96,7 @@ const std::string GetBootstrapModule(const std::vector<std::string> scriptModule
  * - If the `sys` module cannot be imported, an error is logged.
  * - If the `sys.path` attribute cannot be accessed, no action is taken.
  */
-const void AppendSysPathModules(std::vector<std::filesystem::path> sitePackages) 
+MILLENNIUM const void AppendSysPathModules(std::vector<std::filesystem::path> sitePackages) 
 {
     PyObject *sysModule = PyImport_ImportModule("sys");
     if (!sysModule) 
@@ -136,7 +137,7 @@ const void AppendSysPathModules(std::vector<std::filesystem::path> sitePackages)
  * - If the `site` module cannot be imported, the error is printed and logged.
  * - If the `addsitedir` function cannot be retrieved or called, an error is printed and logged.
  */
-void AddSitePackagesDirectory(std::filesystem::path customPath)
+MILLENNIUM void AddSitePackagesDirectory(std::filesystem::path customPath)
 {
     PyObject *siteModule = PyImport_ImportModule("site");
 
@@ -169,7 +170,7 @@ void AddSitePackagesDirectory(std::filesystem::path customPath)
  * 
  * @param global_dict The global dictionary of the Python interpreter.
  */
-void StartPluginBackend(PyObject* global_dict, std::string pluginName) 
+MILLENNIUM void StartPluginBackend(PyObject* global_dict, std::string pluginName) 
 {
     const auto PrintError = [&pluginName]() 
     {
@@ -236,7 +237,7 @@ void StartPluginBackend(PyObject* global_dict, std::string pluginName)
  * - If the `__builtins__` dictionary cannot be retrieved, a `RuntimeError` is raised.
  * - If setting the `MILLENNIUM_PLUGIN_SECRET_NAME` in `__builtins__` fails, a `RuntimeError` is raised.
  */
-const void SetPluginSecretName(PyObject* globalDictionary, const std::string& pluginName) 
+MILLENNIUM const void SetPluginSecretName(PyObject* globalDictionary, const std::string& pluginName) 
 {
     /** Set the secret name in the global dictionary, i.e the global scope */
     PyDict_SetItemString(globalDictionary, "MILLENNIUM_PLUGIN_SECRET_NAME", PyUnicode_FromString(pluginName.c_str()));
@@ -266,7 +267,7 @@ const void SetPluginSecretName(PyObject* globalDictionary, const std::string& pl
  *
  * Both paths are converted to strings and set as Python variables in the global dictionary.
  */
-const void SetPluginEnvironmentVariables(PyObject* globalDictionary, const SettingsStore::PluginTypeSchema& plugin) 
+MILLENNIUM const void SetPluginEnvironmentVariables(PyObject* globalDictionary, const SettingsStore::PluginTypeSchema& plugin) 
 {
     PyDict_SetItemString(globalDictionary, "PLUGIN_BASE_DIR", PyUnicode_FromString(plugin.pluginBaseDirectory.generic_string().c_str()));
     PyDict_SetItemString(globalDictionary, "__file__", PyUnicode_FromString((plugin.backendAbsoluteDirectory / "main.py").generic_string().c_str()));
@@ -289,7 +290,7 @@ const void SetPluginEnvironmentVariables(PyObject* globalDictionary, const Setti
  * Error Handling:
  * - If any step of the process fails (e.g., file opening, module import), the error is logged and the backend load is marked as failed.
  */
-const void CoInitializer::BackendStartCallback(SettingsStore::PluginTypeSchema plugin) 
+MILLENNIUM const void CoInitializer::BackendStartCallback(SettingsStore::PluginTypeSchema plugin) 
 {
     PyObject* globalDictionary = PyModule_GetDict(PyImport_AddModule("__main__"));
     const auto backendMainModule = plugin.backendAbsoluteDirectory.generic_string();
@@ -317,7 +318,7 @@ const void CoInitializer::BackendStartCallback(SettingsStore::PluginTypeSchema p
     CoInitializer::BackendCallbacks& backendHandler = CoInitializer::BackendCallbacks::getInstance();
 
     PyObject *mainModuleObj = Py_BuildValue("s", backendMainModule.c_str());
-    FILE *mainModuleFilePtr = _Py_fopen_obj(mainModuleObj, "r+");
+    FILE *mainModuleFilePtr = _Py_fopen_obj(mainModuleObj, "r");
 
     if (mainModuleFilePtr == NULL) 
     {
@@ -340,6 +341,8 @@ const void CoInitializer::BackendStartCallback(SettingsStore::PluginTypeSchema p
         fclose(mainModuleFilePtr);
         return;
     }
+
+    Logger.Log("Running plugin: {}", plugin.pluginName);
 
     PyObject* result = PyRun_File(mainModuleFilePtr, backendMainModule.c_str(), Py_file_input, mainModuleDict, mainModuleDict);
     fclose(mainModuleFilePtr);
@@ -406,7 +409,7 @@ const void CoInitializer::BackendStartCallback(SettingsStore::PluginTypeSchema p
  *
  * The constructed module is returned as a string.
  */
-const std::string ConstructOnLoadModule(uint16_t ftpPort, uint16_t ipcPort) 
+MILLENNIUM const std::string ConstructOnLoadModule(uint16_t ftpPort, uint16_t ipcPort) 
 {
     std::unique_ptr<SettingsStore> settingsStore = std::make_unique<SettingsStore>();
     std::vector<SettingsStore::PluginTypeSchema> plugins = settingsStore->ParseAllPlugins();
@@ -433,12 +436,12 @@ const std::string ConstructOnLoadModule(uint16_t ftpPort, uint16_t ipcPort)
  * 
  * @note this function is only applicable to Windows
  */
-const void UnPatchSharedJSContext()
+MILLENNIUM const void UnPatchSharedJSContext()
 {
     #ifdef _WIN32
     Logger.Log("Restoring SharedJSContext...");
 
-    const auto SteamUIModulePath = SystemIO::GetSteamPath() / "steamui" / "index.html";
+    const auto SteamUIModulePath       = SystemIO::GetSteamPath() / "steamui" / "index.html";
     const auto SteamUIModulePathBackup = SystemIO::GetSteamPath() / "steamui" / "orig.html";
 
     try
@@ -461,7 +464,6 @@ const void UnPatchSharedJSContext()
 
     Logger.Log("Restored SharedJSContext...");
     #endif
-    // Sockets::PostShared({ { "id", 9773 }, { "method", "Page.reload" } });
 }
 
 
@@ -484,7 +486,7 @@ const void UnPatchSharedJSContext()
  * Error Handling:
  * - If any issues occur during the message processing, errors are logged with details.
  */
-void OnBackendLoad(uint16_t ftpPort, uint16_t ipcPort)
+MILLENNIUM void OnBackendLoad(uint16_t ftpPort, uint16_t ipcPort)
 {
     UnPatchSharedJSContext(); // Restore the original SharedJSContext
     Logger.Log("Notifying frontend of backend load...");
@@ -517,8 +519,9 @@ void OnBackendLoad(uint16_t ftpPort, uint16_t ipcPort)
             {
                 addedScriptOnNewDocumentId = eventMessage["result"]["identifier"];
                 hasScriptIdentifier = true;
-                Logger.Log("Successfully injected shims, updating state...");
-                Sockets::PostShared({ {"id", PAGE_RELOAD }, {"method", "Page.reload"} });
+                Logger.Log("Successfully injected shims, crashing GPU process...");
+
+                Sockets::PostGlobal({ {"id", PAGE_RELOAD }, {"method", "Page.reload"} });
                 cvScript.notify_one();  
 
                 Logger.Log("Successfully notified frontend...");
@@ -543,13 +546,13 @@ void OnBackendLoad(uint16_t ftpPort, uint16_t ipcPort)
     Logger.Log("Frontend notifier finished!");
 }
 
-const void CoInitializer::InjectFrontendShims(uint16_t ftpPort, uint16_t ipcPort) 
+MILLENNIUM const void CoInitializer::InjectFrontendShims(uint16_t ftpPort, uint16_t ipcPort) 
 {
     BackendCallbacks& backendHandler = BackendCallbacks::getInstance();
     backendHandler.RegisterForLoad(std::bind(OnBackendLoad, ftpPort, ipcPort));
 }
 
-const void CoInitializer::ReInjectFrontendShims(std::shared_ptr<PluginLoader> pluginLoader)
+MILLENNIUM const void CoInitializer::ReInjectFrontendShims(std::shared_ptr<PluginLoader> pluginLoader)
 {
     pluginLoader->InjectWebkitShims();