+++ src/sys/env.cc
@@ -41,6 +41,7 @@
 #include <log.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include "fvisible.h"
 
 const void SetupEnvironmentVariables();